requirements.txt
streamlit
pyTelegramBotAPI
openai-whisper
pytz